new buffer appears to work for every combination of channel width, device width, dies, channels and planes that I tried. and I tried a lot of them. I still want to change one aspect of how the die locks on a read to maximize concurrency. Otherwise buffering is good to go.

+++ src/Die.h
@@ -21,7 +21,7 @@ namespace NVDSim{
 	                Die(NVDIMM *parent, Logger *l, uint64_t id);
 			void attachToBuffer(Buffer *buff);
 			void receiveFromBuffer(ChannelPacket *busPacket);
-			int isDieBusy(uint plane);
+			int isDieBusy(uint64_t plane);
 			void update(void);
 			void channelDone(void);
 			void bufferDone(void);
